+++ src/multi_physics/QED/src/physics/breit_wheeler/breit_wheeler_engine_tables_generator.hpp
@@ -84,7 +84,7 @@ namespace breit_wheeler{
 
         int count = 0;
         #pragma omp parallel for
-        for (int i = 0; i < all_coords.size(); ++i){
+        for (int i = 0; i < static_cast<int>(all_coords.size()); ++i){
             PXRMP_INTERNAL_CONSTEXPR_IF (use_internal_double){
                 all_vals[i] = aux_generate_double(all_coords[i]);
             }
